Made Energion Crystal growth rate configurable and faster by default

+++ src/main/java/lykrast/prodigytech/common/block/BlockEnergionCrystal.java
@@ -6,6 +6,7 @@ import javax.annotation.Nullable;
 
 import lykrast.prodigytech.common.init.ModItems;
 import lykrast.prodigytech.common.util.AABBUtil;
+import lykrast.prodigytech.common.util.Config;
 import net.minecraft.block.Block;
 import net.minecraft.block.SoundType;
 import net.minecraft.block.material.Material;
@@ -57,45 +58,31 @@ public class BlockEnergionCrystal extends BlockGeneric implements ICustomStateMa
 
 	//Just like wheat
     @Override
-	public void updateTick(World worldIn, BlockPos pos, IBlockState state, Random rand)
-    {
+	public void updateTick(World worldIn, BlockPos pos, IBlockState state, Random rand) {
         super.updateTick(worldIn, pos, state, rand);
         
         int age = getAge(state);
         
-        if (age < 5)
-        {
-            float f = getGrowthChance(worldIn, pos);
-
-            if(net.minecraftforge.common.ForgeHooks.onCropsGrowPre(worldIn, pos, state, rand.nextInt((int)(25.0F / f) + 1) == 0))
-            {
+        if (age < 5) {
+            if(net.minecraftforge.common.ForgeHooks.onCropsGrowPre(worldIn, pos, state, rand.nextInt((int)(25 / Config.energionGrowthSpeed) + 1) == 0)) {
                 worldIn.setBlockState(pos, this.withAge(age + 1), 2);
                 net.minecraftforge.common.ForgeHooks.onCropsGrowPost(worldIn, pos, state, worldIn.getBlockState(pos));
             }
         }
     }
     
-    protected float getGrowthChance(World worldIn, BlockPos pos)
-    {
-    	//TODO: something more sophisticated
-    	return 1.0f;
-    }
-    
     @Override
-    public Item getItemDropped(IBlockState state, Random rand, int fortune)
-    {
+    public Item getItemDropped(IBlockState state, Random rand, int fortune) {
         return ModItems.energionCrystalSeed;
     }
     
     @Override
-    public ItemStack getPickBlock(IBlockState state, RayTraceResult target, World world, BlockPos pos, EntityPlayer player)
-    {
+    public ItemStack getPickBlock(IBlockState state, RayTraceResult target, World world, BlockPos pos, EntityPlayer player) {
         return new ItemStack(ModItems.energionCrystalSeed);
     }
 
     @Override
-    public void getDrops(NonNullList<ItemStack> drops, IBlockAccess world, BlockPos pos, IBlockState state, int fortune)
-    {
+    public void getDrops(NonNullList<ItemStack> drops, IBlockAccess world, BlockPos pos, IBlockState state, int fortune) {
         int age = getAge(state);
         Random rand = world instanceof World ? ((World)world).rand : new Random();
 
@@ -117,15 +104,13 @@ public class BlockEnergionCrystal extends BlockGeneric implements ICustomStateMa
     }
 
 	@Override
-    public AxisAlignedBB getBoundingBox(IBlockState state, IBlockAccess source, BlockPos pos)
-    {
+    public AxisAlignedBB getBoundingBox(IBlockState state, IBlockAccess source, BlockPos pos) {
         return AABB[((Integer)state.getValue(AGE)).intValue()];
     }
 
     @Nullable
 	@Override
-    public AxisAlignedBB getCollisionBoundingBox(IBlockState state, IBlockAccess worldIn, BlockPos pos)
-    {
+    public AxisAlignedBB getCollisionBoundingBox(IBlockState state, IBlockAccess worldIn, BlockPos pos) {
         return AABB[((Integer)state.getValue(AGE)).intValue()];
     }
 
@@ -133,70 +118,55 @@ public class BlockEnergionCrystal extends BlockGeneric implements ICustomStateMa
      * Checks if this block can be placed exactly at the given position.
      */
 	@Override
-    public boolean canPlaceBlockAt(World worldIn, BlockPos pos)
-    {
+    public boolean canPlaceBlockAt(World worldIn, BlockPos pos) {
         return worldIn.getBlockState(pos.down()).isSideSolid(worldIn, pos.down(), EnumFacing.UP);
     }
 
 	@Override
-    public boolean isOpaqueCube(IBlockState state)
-    {
+    public boolean isOpaqueCube(IBlockState state) {
         return false;
     }
 
 	@Override
-    public boolean isFullCube(IBlockState state)
-    {
+    public boolean isFullCube(IBlockState state) {
         return false;
     }
 
     @SideOnly(Side.CLIENT)
     @Override
-    public BlockRenderLayer getRenderLayer()
-    {
+    public BlockRenderLayer getRenderLayer() {
         return BlockRenderLayer.CUTOUT;
     }
 
 	@Override
-    public BlockFaceShape getBlockFaceShape(IBlockAccess worldIn, IBlockState state, BlockPos pos, EnumFacing face)
-    {
+    public BlockFaceShape getBlockFaceShape(IBlockAccess worldIn, IBlockState state, BlockPos pos, EnumFacing face) {
         return BlockFaceShape.UNDEFINED;
     }
 	
 	@Override
-    public void neighborChanged(IBlockState state, World worldIn, BlockPos pos, Block blockIn, BlockPos fromPos)
-    {
-        if (!this.canPlaceBlockAt(worldIn, pos))
-        {
-            worldIn.destroyBlock(pos, false);
-        }
+    public void neighborChanged(IBlockState state, World worldIn, BlockPos pos, Block blockIn, BlockPos fromPos) {
+        if (!this.canPlaceBlockAt(worldIn, pos)) worldIn.destroyBlock(pos, false);
     }
 	
 	@Override
-    public void onBlockExploded(World worldIn, BlockPos pos, Explosion explosionIn)
-    {
+    public void onBlockExploded(World worldIn, BlockPos pos, Explosion explosionIn) {
 		explode(worldIn, pos, worldIn.getBlockState(pos), explosionIn.getExplosivePlacedBy());        
         super.onBlockExploded(worldIn, pos, explosionIn);
     }
     
     @Override
-    public boolean removedByPlayer(IBlockState state, World world, BlockPos pos, EntityPlayer player, boolean willHarvest)
-    {
+    public boolean removedByPlayer(IBlockState state, World world, BlockPos pos, EntityPlayer player, boolean willHarvest) {
     	boolean destroyed = super.removedByPlayer(state, world, pos, player, willHarvest);
     	if (destroyed && !player.isCreative()) explode(world, pos, state, player);
     	return destroyed;
     }
     
-	private void explode(World world, BlockPos pos, IBlockState state, EntityLivingBase igniter)
-	{
-        if (!world.isRemote)
-        {
+	private void explode(World world, BlockPos pos, IBlockState state, EntityLivingBase igniter) {
+        if (!world.isRemote) {
 	    	int age = getAge(state);
 	    	
-	    	if (!isDefused(state) && age > 0)
-	    	{
-	    		for (int i=0; i<age; i++)
-	    		{
+	    	if (!isDefused(state) && age > 0) {
+	    		for (int i=0; i<age; i++) {
 	                EntityTNTPrimed entitytntprimed = new EntityTNTPrimed(world, (double)((float)pos.getX() + 0.5F), (double)pos.getY(), (double)((float)pos.getZ() + 0.5F), igniter);
 	                entitytntprimed.setFuse(5 + 4*i);
 	                world.spawnEntity(entitytntprimed);
@@ -206,8 +176,7 @@ public class BlockEnergionCrystal extends BlockGeneric implements ICustomStateMa
 	}
 
 	@Override
-    public boolean canDropFromExplosion(Explosion explosionIn)
-    {
+    public boolean canDropFromExplosion(Explosion explosionIn) {
         return false;
     }
 
@@ -217,18 +186,15 @@ public class BlockEnergionCrystal extends BlockGeneric implements ICustomStateMa
 	}
 
 	//Copied from crops
-    public IBlockState withAge(int age)
-    {
+    public IBlockState withAge(int age) {
         return this.getDefaultState().withProperty(AGE, Integer.valueOf(age)).withProperty(DEFUSED, false);
     }
 
-    public static int getAge(IBlockState state)
-    {
+    public static int getAge(IBlockState state) {
         return ((Integer)state.getValue(AGE)).intValue();
     }
     
-    private boolean isDefused(IBlockState state)
-    {
+    private boolean isDefused(IBlockState state) {
     	return state.getValue(DEFUSED);
     }
     
@@ -237,22 +203,19 @@ public class BlockEnergionCrystal extends BlockGeneric implements ICustomStateMa
     }
 
 	@Override
-    public IBlockState getStateFromMeta(int meta)
-    {
+    public IBlockState getStateFromMeta(int meta) {
         return withAge(meta);
     }
 
 	@Override
-    public int getMetaFromState(IBlockState state)
-    {
-        return getAge(state);
-    }
+	public int getMetaFromState(IBlockState state) {
+		return getAge(state);
+	}
 
 	@Override
-    protected BlockStateContainer createBlockState()
-    {
-        return new BlockStateContainer(this, AGE, DEFUSED);
-    }
+	protected BlockStateContainer createBlockState() {
+		return new BlockStateContainer(this, AGE, DEFUSED);
+	}
 
 	@Override
 	@SideOnly(Side.CLIENT)
